RobotTest: use templating instead of abstract class
This removes UdpInterface and instead this is specified as a parameter
to the templated class PcInterface. This enables either a
MockUdpInterface
or the real UDP interface to be specified to the PcInterface object
at compile time.

An undesirable implication of templating is that the template class
member function definitions must exist in the same compilation unit,
i.e. the PcInterface.h file. Otherwise, the linker is unable to locate
the function definitions for e.g. PcInterface_test.cpp to use.